requirements.txt
pygame
pyspellchecker